+++ artifacts/csv_reader.py
@@ -1,0 +1,5 @@
+import pandas as pd
+
+def read_csv_safe(filepath, **kwargs):
+    '''安全读取CSV文件'''
+    return pd.read_csv(filepath, **kwargs)